renamed ensemble strategies
