+++ src/schnetpack/interfaces/__init__.py
@@ -1,3 +1,3 @@
 from .ase_interface import *
 from .batchwise_optimization import *
-from .ensemble_calculator import *
+from .ensemble_strategies import *
